+++ src/api/json.rs
@@ -4,8 +4,10 @@
 use crate::func::native::locked_write;
 use crate::parser::{ParseSettingFlags, ParseState};
 use crate::tokenizer::{lex_raw, Token};
+use crate::types::dynamic::Union;
 use crate::types::StringsInterner;
-use crate::{Engine, LexError, Map, RhaiResultOf};
+use crate::{Dynamic, Engine, LexError, Map, RhaiResultOf};
+use std::fmt::Write;
 #[cfg(feature = "no_std")]
 use std::prelude::v1::*;
 
@@ -166,8 +168,6 @@ pub fn format_map_as_json(map: &Map) -> String {
     let mut result = String::from('{');
 
     for (key, value) in map {
-        use std::fmt::Write;
-
         if result.len() > 1 {
             result.push(',');
         }
@@ -175,14 +175,53 @@ pub fn format_map_as_json(map: &Map) -> String {
         write!(result, "{key:?}").unwrap();
         result.push(':');
 
-        match value.read_lock::<Map>() {
-            Some(val) => result.push_str(&format_map_as_json(&val)),
-            None if value.is_unit() => result.push_str("null"),
-            None => write!(result, "{value:?}").unwrap(),
-        }
+        format_dynamic_as_json(&mut result, value);
     }
 
     result.push('}');
 
     result
+}
+
+/// Format a [`Dynamic`] value as JSON.
+fn format_dynamic_as_json(result: &mut String, value: &Dynamic) {
+    match value.0 {
+        Union::Unit(..) => result.push_str("null"),
+        Union::FnPtr(ref f, _, _) if f.is_curried() => {
+            result.push('[');
+            write!(result, "{:?}", f.fn_name()).unwrap();
+            f.iter_curry().for_each(|value| {
+                result.push(',');
+                format_dynamic_as_json(result, value);
+            });
+            result.push(']');
+        }
+        Union::FnPtr(ref f, _, _) => write!(result, "{:?}", f.fn_name()).unwrap(),
+        Union::Map(ref m, ..) => result.push_str(&format_map_as_json(&m)),
+        #[cfg(not(feature = "no_index"))]
+        Union::Array(ref a, _, _) => {
+            result.push('[');
+            for (i, x) in a.iter().enumerate() {
+                if i > 0 {
+                    result.push(',');
+                }
+                format_dynamic_as_json(result, x);
+            }
+            result.push(']');
+        }
+        #[cfg(not(feature = "no_index"))]
+        Union::Blob(ref b, _, _) => {
+            result.push('[');
+            for (i, x) in b.iter().enumerate() {
+                if i > 0 {
+                    result.push(',');
+                }
+                write!(result, "{x}").unwrap();
+            }
+            result.push(']');
+        }
+        #[cfg(not(feature = "no_closure"))]
+        Union::Shared(ref v, _, _) => format_dynamic_as_json(result, &*v.borrow()),
+        _ => write!(result, "{value:?}").unwrap(),
+    }
 }