+++ src/serde/serialize.rs
@@ -3,6 +3,7 @@
 use crate::types::dynamic::Union;
 use crate::{Dynamic, ImmutableString, Scope};
 use serde::{ser::SerializeSeq, Serialize, Serializer};
+use std::iter::once;
 #[cfg(feature = "no_std")]
 use std::prelude::v1::*;
 
@@ -62,6 +63,9 @@ impl Serialize for Dynamic {
                 let mut map = ser.serialize_map(Some(m.len()))?;
                 m.iter().try_for_each(|(k, v)| map.serialize_entry(k, v))?;
                 map.end()
+            }
+            Union::FnPtr(ref f, ..) if f.is_curried() => {
+                ser.collect_seq(once(f.fn_name().into()).chain(f.iter_curry().cloned()))
             }
             Union::FnPtr(ref f, ..) => ser.serialize_str(f.fn_name()),
             #[cfg(not(feature = "no_time"))]